Move aws iot logic to its own package
